Rename OGL_Renderer files to RendererOpenGL.
