+++ proj/vs2019/NAS2D.vcxproj.filters
@@ -84,9 +84,6 @@
     <ClCompile Include="..\..\src\Resources\Sprite.cpp">
       <Filter>Source Files\Resources</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\Renderer\OGL_Renderer.cpp">
-      <Filter>Source Files\Renderer</Filter>
-    </ClCompile>
     <ClCompile Include="..\..\src\Renderer\Primitives.cpp">
       <Filter>Source Files\Renderer</Filter>
     </ClCompile>
@@ -140,6 +137,9 @@
     </ClCompile>
     <ClCompile Include="..\..\src\Mixer\NullMixer.cpp">
       <Filter>Source Files\Mixer</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\Renderer\RendererOpenGL.cpp">
+      <Filter>Source Files\Renderer</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -215,9 +215,6 @@
     <ClInclude Include="..\..\include\NAS2D\Resources\Sprite.h">
       <Filter>Header Files\Resources</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\include\NAS2D\Renderer\OGL_Renderer.h">
-      <Filter>Header Files\Renderer</Filter>
-    </ClInclude>
     <ClInclude Include="..\..\include\NAS2D\Renderer\RendererNull.h">
       <Filter>Header Files\Renderer</Filter>
     </ClInclude>
@@ -280,6 +277,9 @@
     </ClInclude>
     <ClInclude Include="..\..\include\NAS2D\Mixer\NullMixer.h">
       <Filter>Header Files\Mixer</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\include\NAS2D\Renderer\RendererOpenGL.h">
+      <Filter>Header Files\Renderer</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
